Rework the build/deploy section
I moved around a bunch of text and made sure not to mix the features of
`build` into the sections about `deploy` (even though functionally that
makes sense).
Also linked to the cli reference at the end of each part (since that
documents the options in much greater detail).

Change-type: patch
Connects-to: https://github.com/balena-io/docs/issues/550
Connects-to: https://github.com/balena-io/balena-cli/pull/2272
Signed-off-by: Robert Günzler <robertg@balena.io>